requirements.txt
anyio==4.4.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
contourpy==1.3.0
crossfire==1.1.0
cycler==0.12.1
exceptiongroup==1.2.2
fonttools==4.53.1
geopandas==1.0.1
h11==0.14.0
httpcore==1.0.5
httpx==0.25.2
idna==3.8
joblib==1.4.2
kiwisolver==1.4.5
libpysal==4.12.1
matplotlib==3.9.2
nest-asyncio==1.6.0
numpy==2.1.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
pointpats==2.5.0
pyarrow==17.0.0
pyogrio==0.9.0
pyparsing==3.1.4
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
scikit-learn==1.5.1
scipy==1.14.1
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
threadpoolctl==3.5.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
